IOT-Presentation-Blink-LED-when-number-of-comments-reached-on-last-post/iot-led-blinking/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dnspython==1.16.0
dpath==1.4.2
eventlet==0.24.1
Flask==1.0.2
Flask-SocketIO==3.3.2
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.3.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
monotonic==1.5
multidict==4.5.2
pyserial==3.4
python-engineio==3.5.0
python-socketio==3.1.2
requests==2.21.0
six==1.12.0
socketIO-client==0.7.2
SQLAlchemy==1.3.1
typing-extensions==3.7.2
urllib3==1.24.1
websocket-client==0.55.0
websockets==7.0
Werkzeug==0.14.1
yarl==1.3.0
